requirements.txt
flask
tensorflow
numpy
Pillow
joblib
ollama
